+++ src/main/java/com/powsybl/openloadflow/ac/TransformerVoltageControlOuterLoop.java
@@ -7,19 +7,17 @@
 package com.powsybl.openloadflow.ac;
 
 import com.powsybl.commons.reporter.Reporter;
-import com.powsybl.openloadflow.ac.outerloop.OuterLoop;
 import com.powsybl.openloadflow.ac.outerloop.OuterLoopContext;
 import com.powsybl.openloadflow.ac.outerloop.OuterLoopStatus;
-import com.powsybl.openloadflow.network.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.powsybl.openloadflow.network.LfBranch;
+import com.powsybl.openloadflow.network.LfBus;
+import com.powsybl.openloadflow.network.LfNetwork;
+import com.powsybl.openloadflow.network.TransformerVoltageControl;
 
 /**
  * @author Anne Tilloy <anne.tilloy at rte-france.com>
  */
-public class TransformerVoltageControlOuterLoop implements OuterLoop {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(TransformerVoltageControlOuterLoop.class);
+public class TransformerVoltageControlOuterLoop extends AbstractTransformerVoltageControlOuterLoop {
 
     private double maxControlledNominalVoltage = Double.MIN_VALUE;
 
@@ -70,20 +68,7 @@ public class TransformerVoltageControlOuterLoop implements OuterLoop {
         // At second outer loop iteration, the transformers are rounded. The generator voltage controls that were
         // disabled previously are enabled.
         if (context.getIteration() == 1) {
-            for (LfBranch branch : context.getNetwork().getBranches()) {
-                TransformerVoltageControl voltageControl = branch.getVoltageControl().orElse(null);
-                if (voltageControl != null) {
-                    branch.setVoltageControlEnabled(false);
-
-                    // round the rho shift to the closest tap
-                    PiModel piModel = branch.getPiModel();
-                    double r1Value = piModel.getR1();
-                    piModel.roundR1ToClosestTap();
-                    double roundedR1Value = piModel.getR1();
-                    LOGGER.trace("Round voltage ratio of '{}': {} -> {}", branch.getId(), r1Value, roundedR1Value);
-                    status = OuterLoopStatus.UNSTABLE;
-                }
-            }
+            status = roundVoltageRatios(context.getNetwork());
             for (LfBus bus : context.getNetwork().getBuses()) {
                 if (bus.hasVoltageControllerCapability() && bus.getNominalV() <= maxControlledNominalVoltage) {
                     bus.setGenerationTargetQ(0);
@@ -94,12 +79,5 @@ public class TransformerVoltageControlOuterLoop implements OuterLoop {
         }
 
         return status;
-    }
-
-    @Override
-    public void cleanup(LfNetwork network) {
-        for (LfBranch branch : network.getBranches()) {
-            branch.getVoltageControl().ifPresent(voltageControl -> branch.setVoltageControlEnabled(true));
-        }
     }
 }
